refactor(ssp): 重构广告位聚合实体和索引服务
- 引入 SiteAdPlacementAggregate 聚合实体替代原有 SiteAdPlacement
- 重构缓存配置和服务类以支持新的聚合实体类型
- 更新索引服务实现使用统一的索引键构建逻辑
- 添加诊断控制器用于构建索引键调试功能
- 优化应用预热服务使用虚拟线程异步初始化
- 在广告拉取接口中添加异常处理和日志记录
- 更新测试类配置以支持新的依赖注入方式
- 修改数据访问层方法参数类型保持一致性

+++ open-adexchange-mos/src/main/java/top/openadexchange/mos/api/SiteAdPlacementController.java
@@ -52,7 +52,7 @@ public class SiteAdPlacementController {
 
     @GetMapping("/{id}")
     @Operation(summary = "获取媒体广告位")
-    public ApiResponse<SiteAdPlacementDto> getSiteAdPlacement(@PathVariable("id") Long id) {
+    public ApiResponse<SiteAdPlacementDto> getSiteAdPlacement(@PathVariable("id") Integer id) {
         return ApiResponse.success(siteAdPlacementService.getSiteAdPlacement(id));
     }
